+++ src/bittytax/conv/parsers/__init__.py
@@ -69,6 +69,7 @@ from . import (
     poloniex,
     qtrade,
     qtwallet,
+    robinhood,
     snowtrace,
     staketax,
     subscan,