+++ CHANGELOG.md
@@ -122,6 +122,7 @@
 - Conversion tool: added parser for Voyager exchange.
 - Conversion tool: added parser for LBank exchange.
 - Conversion tool: added parser for Bybit exchange.
+- Conversion tool: added parser for Robinhood exchange.
 ### Changed
 - Conversion tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))
 - Accounting tool: openpyxl use read-only mode. ([#337](https://github.com/BittyTax/BittyTax/issues/337))